changes 4 Prop Modules  and prop  users *And Version Change

+++ ERPWebAPI.EL/Concrete/HR/HR_tbl_VacationGroupContent.cs
@@ -7,8 +7,8 @@ namespace ERPWebAPI.EL.Concrete.HR
     {
         [Key]
         public short ContentId { get; set; }
-        public short VacationGroupId { get; set; }
-        public byte VacationTypeId { get; set; }
+        public byte VacationGroupId { get; set; }
+        public short VacationTypeId { get; set; }
         public string? VacationType { get; set; }
         public string? UserEmployee { get; set; }
         public DateTime TransactionDate { get; set; }
